multi-agent/requirements.txt
# Core dependencies
anthropic>=0.68.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
aiofiles>=23.0.0

# Compatibility fixes
numpy>=1.24.0,<2.0  # Avoid NumPy 2.x compatibility issues
protobuf>=4.25.3,<5.0  # Fix MessageFactory issues

# Document processing
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0
tiktoken>=0.5.0

# Vector database and embeddings (compatible versions)
qdrant-client>=1.7.0
chromadb>=0.4.0,<1.1.0  # Pin to avoid conflicts
sentence-transformers>=2.3.0,<3.0  # Avoid NumPy 2.x issues
faiss-cpu>=1.7.4

# Search and retrieval
rank-bm25>=0.2.2
scikit-learn>=1.3.0

# Progress and monitoring
tqdm>=4.66.0

# Optional: Advanced features (uncomment if needed)
# redis>=5.0.0              # For caching
# celery>=5.3.0             # For distributed processing
# prometheus-client>=0.19.0 # For metrics
# torch>=2.0.0              # GPU acceleration (CPU-only by default)